requirements.txt
# ==== Core ====
numpy>=1.23
scipy>=1.10
PyYAML>=6.0

# Medical imaging
nibabel>=5.2
SimpleITK>=2.3
pydicom>=2.4

# Image processing & viz
scikit-image>=0.24
opencv-python-headless>=4.10
matplotlib>=3.5

# Deep learning (match your CUDA / OS)
torch>=2.2           # install the CUDA build appropriate for your GPU
# torchvision/torchaudio only if you actually use them:
# torchvision>=0.17
# torchaudio>=2.2

# Optional frameworks/utilities
monai>=1.3           # helpful for 3D aug/infer; optional
tqdm>=4.66
tensorboard>=2.17
einops>=0.8          # only needed if you use it in models (safe to keep)

# (optional) stronger 3D aug if you decide to use it:
# torchio>=0.19
